MapCanvasMapを別ファイルに独立、MapCanvasMapにshoudComponentUpdateを追加
FIXME: パーツを配置しても配置したパーツがマップに反映されない
