+++ packages/maker/src/common/MapCanvas.tsx
@@ -1,18 +1,7 @@
-import React, { RefObject, useMemo, ReactNode } from 'react';
-import { useSelector } from 'react-redux';
+import React, { RefObject, ReactNode } from 'react';
 import WWAConsts from '../classes/WWAConsts';
-import { Coord } from '@wwawing/common-interface';
 import styles from './MapCanvas.module.scss';
-import { StoreType } from '../State';
-
-/**
- * 一番下に敷かれる背景色です。
- */
-const MAP_CANVAS_BASE_COLOR = '#000';
-/**
- * MapChunk のサイズです。一番左のチャンクは横幅1マス、一番上のチャンクは縦幅1マス増えます。
- */
-const CHUNK_SIZE = 20;
+import MapCanvasMap from './MapCanvasMap';
 
 type MouseEventFunc = (x: number, y: number) => void;
 
@@ -136,180 +125,6 @@ export class MapCanvas extends React.Component<Props, State> {
         );
     }
 }
-
-const MapCanvasMap: React.FC = () => {
-    const fieldMap = useSelector((state: StoreType) => {
-        if (state.wwaData === null) {
-            return [];
-        }
-
-        return [
-            {
-                map: state.wwaData.map,
-                attribute: state.wwaData.mapAttribute
-            }, {
-                map: state.wwaData.mapObject,
-                attribute: state.wwaData.objectAttribute
-            }
-        ];
-    });
-    const image = useSelector((state: StoreType) => state.image) || new Image();
-    const mapWidth = useSelector((state: StoreType) => state.wwaData?.mapWidth);
-
-    const chunks = useMemo(() => {
-        if (mapWidth === undefined) {
-            return [];
-        }
-
-        /**
-         * 1つの Canvas 要素だけでは描画しきれないため、複数のチャンクに分割し、1チャンクに MapChunk コンポーネントを渡します。
-         */
-        let map: Coord[][][][][] = []; // チャンクY, チャンクX, レイヤー, マスY, マスX
-        fieldMap.forEach((layer, layerIndex) => {
-
-            let screenY = 0;
-            for (let y = 1; y < mapWidth; y += CHUNK_SIZE) {
-
-                map[screenY] = layerIndex === 0 ? [] : map[screenY];
-                const startSliceY = y === 1 ? 0 : y;
-                const endSliceY = y + CHUNK_SIZE;
-
-                let screenX = 0;
-                for (let x = 1; x < mapWidth; x += CHUNK_SIZE) {
-
-                    map[screenY][screenX] = layerIndex === 0 ? [] : map[screenY][screenX];
-                    const startSliceX = x === 1 ? 0 : x;
-                    const endSliceX = x + CHUNK_SIZE;
-
-                    const targetMap = layer.map.slice(startSliceY, endSliceY).map(chunkLine => {
-                        return chunkLine.slice(startSliceX, endSliceX).map(partsNumber => {
-                            return {
-                                x: layer.attribute[partsNumber][WWAConsts.ATR_X],
-                                y: layer.attribute[partsNumber][WWAConsts.ATR_Y]
-                            };
-                        });
-                    });
-
-                    map[screenY][screenX].push(targetMap);
-                    screenX++;
-                }
-
-                screenY++;
-            }
-        });
-
-        return map;
-    }, [fieldMap, mapWidth]);
-
-    return (
-        <>
-            {chunks.map((chunkLine, chunkLineIndex) => (
-                <div className={styles.mapCanvasLine} key={chunkLineIndex}>
-                    {chunkLine.map((chunk, chunkColumnIndex) => (
-                        <MapChunk
-                            key={chunkColumnIndex}
-                            map={chunk}
-                            image={image}
-                        />
-                    ))}
-                </div>
-            ))}
-        </>
-    );
-}
-
-MapCanvasMap.whyDidYouRender = true;
-
-/**
- * 1画面程度の画面を描画する Canvas 要素です。
- *     プロパティ map はイメージ画像で参照される場所の座標を記載した配列になっています。
- */
-class MapChunk extends React.Component<{
-    map: Coord[][][],
-    image: CanvasImageSource
-}> {
-    private canvasRef: React.RefObject<HTMLCanvasElement>;
-
-    constructor(props: MapChunk["props"]) {
-        super(props);
-        this.canvasRef = React.createRef();
-    }
-
-    componentDidMount() {
-        this.draw();
-    }
-
-    componentDidUpdate() {
-        this.draw();
-    }
-
-    /**
-     * 各マスの変更を確認し、変更が見つかった場合にだけ componentDidUpdate を実行させます。
-     */
-    shouldComponentUpdate(prevProps: MapChunk["props"]) {
-        return prevProps.map.some((line, lineIndex) =>
-            line.some((layer, layerIndex) =>
-                layer.some((imageCoord, index) =>
-                    imageCoord.x !== this.props.map[lineIndex][layerIndex][index].x ||
-                    imageCoord.y !== this.props.map[lineIndex][layerIndex][index].y
-                )
-            )
-        );
-    }
-
-    private draw() {
-        if (this.canvasRef.current === null) {
-            return;
-        }
-        const context = this.canvasRef.current.getContext('2d');
-        if (context === null) {
-            return;
-        }
-
-        const elementSize = this.getElementSize();
-        context.fillStyle = MAP_CANVAS_BASE_COLOR;
-        context.fillRect(0, 0, elementSize.x, elementSize.y);
-
-        this.props.map.forEach((layer, index) => {
-            layer.forEach((line, y) => {
-                line.forEach((imageCropCoord, x) => {
-                    if (imageCropCoord.x === 0 && imageCropCoord.y === 0) {
-                        return;
-                    }
-                    context.drawImage(
-                        this.props.image,
-                        imageCropCoord.x,
-                        imageCropCoord.y,
-                        WWAConsts.CHIP_SIZE,
-                        WWAConsts.CHIP_SIZE,
-                        x * WWAConsts.CHIP_SIZE,
-                        y * WWAConsts.CHIP_SIZE,
-                        WWAConsts.CHIP_SIZE,
-                        WWAConsts.CHIP_SIZE
-                    );
-                });
-            });
-        });
-    }
-
-    private getElementSize(): Coord {
-        return {
-            x: this.props.map[0][0].length * WWAConsts.CHIP_SIZE,
-            y: this.props.map[0].length * WWAConsts.CHIP_SIZE
-        };
-    }
-
-    public render() {
-        const elementSize = this.getElementSize();
-        return (
-            <canvas
-                ref={this.canvasRef}
-                width={elementSize.x}
-                height={elementSize.y}
-            />
-        );
-    }
-};
 
 export interface SelectRectProps {
     chipX: number;
